Split characters data into their respective folders
